+++ lib/ullrev.c
@@ -1,12 +1,12 @@
 #include "euler/config.h"
 #include "euler/reverse.h"
 
-ULL
-ullreverse(ULL num)
+u8
+ullreverse(u8 num)
 {
 	int ret = 0;
 
-	while(num)
+	while (num)
 	{
 		ret = ret * 10 + num % 10;
 		num /= 10;
